backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
pgvector-sqlalchemy
sentence-transformers
pandas
requests
aiohttp